+++ dropwizard-assets/src/main/java/com/codahale/dropwizard/assets/AssetsBundle.java
@@ -30,7 +30,7 @@ public class AssetsBundle implements Bundle {
     }
 
     /**
-     * Creates a new AssetsBundle which will configure the service to serve the static files
+     * Creates a new AssetsBundle which will configure the application to serve the static files
      * located in {@code src/main/resources/${path}} as {@code /${path}}. For example, given a
      * {@code path} of {@code "/assets"}, {@code src/main/resources/assets/example.js} would be
      * served up from {@code /assets/example.js}.
@@ -43,7 +43,7 @@ public class AssetsBundle implements Bundle {
     }
 
     /**
-     * Creates a new AssetsBundle which will configure the service to serve the static files
+     * Creates a new AssetsBundle which will configure the application to serve the static files
      * located in {@code src/main/resources/${resourcePath}} as {@code /${uriPath}}. For example, given a
      * {@code resourcePath} of {@code "/assets"} and a uriPath of {@code "/js"},
      * {@code src/main/resources/assets/example.js} would be served up from {@code /js/example.js}.
@@ -57,7 +57,7 @@ public class AssetsBundle implements Bundle {
     }
 
     /**
-     * Creates a new AssetsBundle which will configure the service to serve the static files
+     * Creates a new AssetsBundle which will configure the application to serve the static files
      * located in {@code src/main/resources/${resourcePath}} as {@code /${uriPath}}. If no file name is
      * in ${uriPath}, ${indexFile} is appended before serving. For example, given a
      * {@code resourcePath} of {@code "/assets"} and a uriPath of {@code "/js"},